requirements.txt
pytest
streamlit
langchain
langchain-community
langgraph
langgraph-checkpoint-sqlite
pydantic
python-dotenv
pandas
openpyxl
plotly
python-docx
streamlit-plotly-events
